Change back to int

+++ cmd-tf/cmd_tf/dataloader.py
@@ -23,7 +23,7 @@ class XUnetBatchgen(keras.utils.Sequence):
         for j, path in enumerate(batch_input_img_paths):
             img = load_img(path, target_size=self.img_size)
             x[j] = img
-        y = np.zeros((self.batch_size,) + self.img_size, dtype="float32")
+        y = np.zeros((self.batch_size,) + self.img_size, dtype="uint8")
         for j, path in enumerate(batch_target_img_paths):
             y[j] = load_img(path, target_size=self.img_size, color_mode="grayscale")
         return x, y
